+++ SampleCode/StdDriver/WDT_TimeoutReset/main.c
@@ -76,7 +76,7 @@ void SYS_Init(void)
     CLK_EnableModuleClock(WDT_MODULE);    
         
     /* Select WDT module clock source */
-    CLK_SetModuleClock(WDT_MODULE, CLK_CLKSEL1_WDT_S_LIRC, NULL);
+    CLK_SetModuleClock(WDT_MODULE, CLK_CLKSEL1_WDT_S_LIRC, (uint32_t)NULL);
 
     /*---------------------------------------------------------------------------------------------------------*/
     /* Init I/O Multi-function                                                                                 */
