+++ SampleCode/StdDriver/WWDT_CompareINT/main.c
@@ -80,7 +80,7 @@ void SYS_Init(void)
     CLK_EnableModuleClock(WWDT_MODULE);    
         
     /* Select WWDT module clock source */
-    CLK_SetModuleClock(WWDT_MODULE, CLK_CLKSEL2_WWDT_S_HCLK_DIV2048, NULL);
+    CLK_SetModuleClock(WWDT_MODULE, CLK_CLKSEL2_WWDT_S_HCLK_DIV2048, (uint32_t)NULL);
 
     /*---------------------------------------------------------------------------------------------------------*/
     /* Init I/O Multi-function                                                                                 */
